Moved isValidDID to ipfs package
