+++ packages/ipfs/package.json
@@ -14,6 +14,9 @@
     "./kubo": {
       "import": "./dist/kubo-client.js",
       "types": "./dist/types/kubo-client.d.ts"
+    },
+    "./utils": {
+      "import": "./dist/utils.js"
     }
   },
   "scripts": {
